requirements.txt
astroid==2.3.0
autopep8==1.4.4
Click==7.0
Flask==1.0.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.1
six==1.12.0
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
